+++ CHANGELOG.md
@@ -20,6 +20,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - skip price fetching in AmguConsumer when amgu price is zero
 - cancelRequest is no longer amguPayable
 - in Funds, engine and mlnToken are fetched dynamically from registry (no longer in routes)
+- split amguPayable into amguPayable and amguPayableWithIncentive
 
 ### Removed